requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
phonenumbers==8.13.50
phonenumberslite==8.13.50
pillow==11.0.0
sqlparse==0.5.2
